requirements.txt.txt
fastapi
uvicorn
jinja2
transformers
sentence-transformers
faiss-cpu
torch
sentencepiece
python-multipart
